moved side subscription banner to single component
