+++ src/pages/academy/blog/posts/preview/index.js
@@ -28,7 +28,7 @@ import {
     StyledBreadcrumbsTitle,
 } from '../_style'
 import Banner from '../../../../blog/components/_banner'
-import ArticleEmailBanner from '../../components/side-subscription-banner'
+import ArticleEmailBanner from '../../components/_side-subscription-banner'
 import SocialSharing from '../../../../blog/_social-sharing'
 import { localize, WithIntl } from 'components/localization'
 import Layout from 'components/layout/layout'
